+++ library-valcon-internship/src/components/Modal/Modal.css
@@ -35,7 +35,7 @@
   flex-direction: column;
   border-radius: 8px;
   align-items: center;
-  width: 25%;
+  width: 30%;
 }
 
 .modal-btns {
